refactor(stage-2b): split the store into slices; explicit persistence init
Per docs/plans/stages/stage-2b-store-slices.md — file org only, the public
store API and flat state shape are unchanged (existing suites pass with
zero edits):
- store/slices/: gameSlice (deal, engine-backed wrappers, import/export),
  uiSlice (selection/validation/toggles), autoplaySlice, replaySlice,
  aiSlice (controller delegations + config), sessionSlice
- store/replayReducer.ts: the record-level board re-derivation extracted
  from goToReplayIndex as a pure, directly-tested function (the one
  sanctioned non-engine board walker)
- store/initialState.ts: deal/hydration/restore helpers
- store/persistence.ts: autosave + tab-anchor wiring as explicit
  initSessionPersistence(store) called from main.tsx — importing the store
  no longer attaches subscribers (tests construct stores side-effect free)
- store/engine.ts: the shared GameEngine instance
- gameStore.ts is now a 44-line assembly; no store source file over 290 lines
- ControlPanel: bare useGameStore() destructure -> narrow selectors
- New tests: replayReducer (initial deal, full-history parity, recycle,
  input immutability), persistence init (single subscription, debounce,
  signature dedupe)

App suite 385 green; session-persistence + user-journey + smoke e2e (17) green.
